Håndtere dager per uke med desimaler (#205)

+++ src/main/kotlin/no/nav/amt/arena/acl/domain/kafka/arena/ArenaDeltaker.kt
@@ -39,7 +39,7 @@ data class ArenaDeltaker(
 	val OPPLYSNINGER_INNSOK: String? = null,
 	val PARTISJON: Int? = null,
 	val BEGRUNNELSE_BESTILLING: String? = null,
-	val ANTALL_DAGER_PR_UKE: Int? = null
+	val ANTALL_DAGER_PR_UKE: Float? = null
 ) {
 
 	private val log = LoggerFactory.getLogger(javaClass)
